Adds basic `Watch` implementation.
This uses the `Watch` API to observe the upgrade memory address and log it whenever it changes. Unfortunately this requires polling at every update to look for changes, which is a bit unfortunate. I did try `IMemoryEventsApi` which is event-driven, however that simply never called back with any informtation. The `Watch` API seems to be the best I can do.

Logging every frame significantly slows down the emulator, so it only logs on change.