Adds `Watcher`.
This converts the polling interface of `Watch` into an event interface.

I realized in this that the `Watch` interface only emits integers, not arbitrary bytes or sizes. This could be an issue when trying to extract information from large memory spaces (like arrays) or when integers are treated as unsigned in the game. I'll deal with that when it comes up though.