Adds `SaveContext`.
This is also a record, however records don't seem to play well with references types when strict null checks are enabled. It wants me to mark the field as nullable, which shouldn't be the case here.

Easiest fix is to use a positional record, which makes the construction a little more awkward, but C# still has named function parameters, so its not that bad.