Refactors data models into a `Common` assembly.
This allows both the extractor form and the server to depend on the same objects. The new assembly targets .NET Standard 2.0 because Standard is compatible with both .NET Framework and .NET Core applications. I tried using .NET Standard 2.1, however it seems that .NET Framework 4.8 does not (and probably never will) support .NET Standard 2.1, so the best we can do is .NET Standard 2.0.

https://devblogs.microsoft.com/dotnet/announcing-net-standard-2-1/#net-platform-support