Adds Angular client app.
Created a new Angular client with `npm install @angular/cli@11.2.7; ng new OotStateExtractorClient`.

I initially tried to use the "New Project" dialog of Visual Studio and pick the Angular app. This worked, but embedding the built Angular files into the BizHawk plugin is a bit more complicated. The build process would need to copy the files to a relevant location in the BizHawk install folder and then the server needs to be started at the relevant location. BizHawk doesn't really have a layout to support this and I'm not super familiar with how to customize the buidl process in Visual Studio.

Arguably, a web app viewer should be independent of the BizHawk plugin anyways, since the plugin is simply an extractor which exposes its information via a REST API. It should be compatible with any client, not just its own (otherwise, why make a web app, just embed the client in the external tool form).

As a result, I decided to just make an Angular app manually via the Angular CLI. This requires a second step to run, rather than relying on Visual Studio's play button, but I think it is a simpler infrastructure and makes more architectural sense.