Adds a `/stream` endpoint to `save-context` which provides server-sent events to the client.
This endlessly streams the current state of the `SaveContext` via server-sent events. This is done by simply serializing the data and then formatting each line to being with `data: ` followed by trailing empty newlines. This is part of the spec and the browser will trims these out to identify a single concrete message for the client.

For now, we simply emit the latest state every 5 seconds. This is basically live updating, just in a very wasteful fashion with an SLO of 5 seconds.