Updates ASP.NET Core dependency injection to treat `EventWithLatest<SaveContext>` as a singleton.
I found that refreshing the client would cause the extractor to stop updating. Eventually I noticed the injected save context was disposed. Breakpointing on the `Dispose()` function shows the ASP.NET dependency injector in the stack trace, leading me to think the server was disposing of the object. The object was provided through an implementation factory, just providing the same object every time. ASP.NET seems to assume that the provided object exists only for the lifetime of its use and automatically calls `Dispose()`, which would break the object for the next request.

Solution is to provide the object as a singleton, which is fairly accurate from the server's point of view. This tells ASP.NET that the same `EventWithLatest<SaveContext>` will be used for all requests, so it should **not** dispose it after a request.